feat: add more code into server core
