+++ packages/server/prod-server-new/src/index.ts
@@ -1,35 +1,24 @@
 import { Server as NodeServer } from 'http';
-import { Hono } from 'hono';
-import { ModernServerOptions } from './type';
-import { Server } from './Server';
+import { ServerCoreOptions } from './type';
+import { ServerCore } from './core';
 import { createNodeServer } from './adapters/node';
-import { NodeRequest, NodeResponse } from './adapters/types';
 
-export default async (
-  options: Omit<ModernServerOptions, 'core'>,
-): Promise<NodeServer> => {
+async function createServer(options: ServerCoreOptions) {
+  const server = new ServerCore(options);
+
+  // server.use(otherMiddlewars )
+
+  await server.init();
+
+  return server;
+}
+
+export default async (options: ServerCoreOptions): Promise<NodeServer> => {
   if (options == null) {
     throw new Error('can not start server without options');
   }
 
-  type Bindings = {
-    node?: {
-      req: NodeRequest;
-      res: NodeResponse;
-    };
-  };
-
-  const hono = new Hono<{ Bindings: Bindings }>();
-
-  const server = new Server({
-    ...options,
-    core: {
-      ...hono,
-      handle: hono.fetch,
-    },
-  });
-
-  await server.init();
+  const server = await createServer(options);
 
   const nodeServer = createNodeServer(server.handle.bind(server));
